// backend/src/types/code/groupCode.ts
export enum GroupErrorCode {
    GROUP_FETCH_FAILED = "GROUP_FETCH_FAILED",
    GROUP_CREATE_FAILED = "GROUP_CREATE_FAILED",
    GROUP_ALREADY_IN_GROUP = "GROUP_ALREADY_IN_GROUP",
    GROUP_NOT_IN_GROUP = "GROUP_NOT_IN_GROUP",
    GROUP_NOT_FOUND = "GROUP_NOT_FOUND",
    GROUP_FORBIDDEN = "GROUP_FORBIDDEN",
    GROUP_INVITE_FAILED = "GROUP_INVITE_FAILED",
    GROUP_INVITE_INVALID = "GROUP_INVITE_INVALID",
    GROUP_JOIN_FAILED = "GROUP_JOIN_FAILED",
    GROUP_LEAVE_FAILED = "GROUP_LEAVE_FAILED",
}